refactor: balances observer ckBTC and Icrc prefixes
