+++ frontend/src/tests/lib/components/accounts/IcrcBalancesObserver.spec.ts
@@ -1,71 +1,158 @@
-import IcrcBalancesObserver from "$lib/components/accounts/IcrcBalancesObserver.svelte";
-import { FETCH_ROOT_KEY, HOST } from "$lib/constants/environment.constants";
-import type { BalancesObserverData } from "$lib/types/icrc.observer";
-import type { PostMessageDataRequestBalances } from "$lib/types/post-message.balances";
-import { mockMainAccount } from "$tests/mocks/icp-accounts.store.mock";
-import { ledgerCanisterIdMock } from "$tests/mocks/sns.api.mock";
+import { CKTESTBTC_UNIVERSE_CANISTER_ID } from "$lib/constants/ckbtc-canister-ids.constants";
+import { AppPath } from "$lib/constants/routes.constants";
+import { icrcAccountsStore } from "$lib/stores/icrc-accounts.store";
+import { syncStore } from "$lib/stores/sync.store";
+import type { Account } from "$lib/types/account";
+import type {
+  PostMessageDataRequestBalances,
+  PostMessageDataResponseBalances,
+} from "$lib/types/post-message.balances";
+import type { PostMessageDataResponseSync } from "$lib/types/post-message.sync";
+import type { PostMessage } from "$lib/types/post-messages";
+import { page } from "$mocks/$app/stores";
+import IcrcBalancesObserverTest from "$tests/lib/components/accounts/IcrcBalancesObserverTest.svelte";
+import { mockCkBTCMainAccount } from "$tests/mocks/ckbtc-accounts.mock";
+import { PostMessageMock } from "$tests/mocks/post-message.mocks";
 import { render, waitFor } from "@testing-library/svelte";
+import { get } from "svelte/store";
 
 describe("IcrcBalancesObserver", () => {
-  let spyPostMessage;
+  type BalancesMessageEvent = MessageEvent<
+    PostMessage<PostMessageDataResponseBalances | PostMessageDataResponseSync>
+  >;
+
+  let postMessageMock: PostMessageMock<BalancesMessageEvent>;
 
   beforeEach(() => {
-    spyPostMessage = vi.fn();
+    const accounts: Account[] = [mockCkBTCMainAccount];
+    icrcAccountsStore.set({
+      universeId: CKTESTBTC_UNIVERSE_CANISTER_ID,
+      accounts: { accounts, certified: true },
+    });
+
+    page.mock({
+      routeId: AppPath.Wallet,
+      data: { universe: CKTESTBTC_UNIVERSE_CANISTER_ID.toText() },
+    });
+
+    postMessageMock = new PostMessageMock();
 
     vi.doMock("$lib/workers/balances.worker?worker", () => ({
       default: class BalancesWorker {
-        postMessage(data: {
+        constructor() {
+          postMessageMock.subscribe(async (msg) => await this.onmessage(msg));
+        }
+
+        postMessage(_data: {
           msg: "nnsStartBalancesTimer" | "nnsStopBalancesTimer";
           data?: PostMessageDataRequestBalances;
         }) {
-          spyPostMessage(data);
+          // Nothing here
         }
+
+        onmessage = async (_params: BalancesMessageEvent) => {
+          // Nothing here
+        };
       },
     }));
   });
 
-  const data: BalancesObserverData = {
-    ledgerCanisterId: ledgerCanisterIdMock.toText(),
-    accounts: [mockMainAccount],
-  };
+  it("should init data and render slotted content", async () => {
+    const { getByTestId } = render(IcrcBalancesObserverTest);
 
-  it("should init worker with parameters", async () => {
-    render(IcrcBalancesObserver, {
-      props: {
-        data,
-        callback: vi.fn(),
-      },
-    });
-
-    await waitFor(() =>
-      expect(spyPostMessage).toBeCalledWith({
-        msg: "nnsStartBalancesTimer",
-        data: {
-          ledgerCanisterId: data.ledgerCanisterId,
-          accountIdentifiers: data.accounts.map(({ identifier }) => identifier),
-          host: HOST,
-          fetchRootKey: FETCH_ROOT_KEY,
-        },
-      })
-    );
+    await waitFor(() => expect(getByTestId("test-observer")).not.toBeNull());
   });
 
-  it("should stop worker on destroy", async () => {
-    const { unmount } = render(IcrcBalancesObserver, {
+  it("should update account store on new sync message", async () => {
+    const { getByTestId } = render(IcrcBalancesObserverTest);
+
+    await waitFor(() => expect(getByTestId("test-observer")).not.toBeNull());
+
+    const accountsInStore = get(icrcAccountsStore);
+    expect(
+      accountsInStore[CKTESTBTC_UNIVERSE_CANISTER_ID.toText()].accounts
+    ).toEqual([mockCkBTCMainAccount]);
+
+    await waitFor(() => expect(postMessageMock.ready).toBeTruthy());
+
+    postMessageMock.emit({
+      data: {
+        msg: "nnsSyncBalances",
+        data: {
+          balances: [
+            {
+              accountIdentifier: mockCkBTCMainAccount.identifier,
+              balance: 123456n,
+            },
+          ],
+        },
+      },
+    } as BalancesMessageEvent);
+
+    await waitFor(() => {
+      const updatedStore = get(icrcAccountsStore);
+      expect(
+        updatedStore[CKTESTBTC_UNIVERSE_CANISTER_ID.toText()].accounts
+      ).toEqual([{ ...mockCkBTCMainAccount, balanceE8s: 123456n }]);
+    });
+  });
+
+  it("should populate error to sync store on error message from worker", async () => {
+    const { getByTestId } = render(IcrcBalancesObserverTest);
+
+    await waitFor(() => expect(getByTestId("test-observer")).not.toBeNull());
+
+    const store = get(syncStore);
+    expect(store).toEqual({
+      balances: "idle",
+      transactions: "idle",
+    });
+
+    await waitFor(() => expect(postMessageMock.ready).toBeTruthy());
+
+    postMessageMock.emit({
+      data: {
+        msg: "nnsSyncErrorBalances",
+      },
+    } as BalancesMessageEvent);
+
+    await waitFor(() => {
+      const updatedStore = get(syncStore);
+      expect(updatedStore).toEqual({
+        balances: "error",
+        transactions: "idle",
+      });
+    });
+  });
+
+  it("should reload on new sync message", async () => {
+    const spyReload = vi.fn();
+    const { getByTestId } = render(IcrcBalancesObserverTest, {
       props: {
-        data,
-        callback: vi.fn(),
+        reload: spyReload,
       },
     });
 
-    await waitFor(() => expect(spyPostMessage).toHaveBeenCalledTimes(1));
+    await waitFor(() => expect(getByTestId("test-observer")).not.toBeNull());
 
-    unmount();
+    expect(spyReload).not.toHaveBeenCalled();
 
-    await waitFor(() => expect(spyPostMessage).toHaveBeenCalledTimes(2));
+    await waitFor(() => expect(postMessageMock.ready).toBeTruthy());
 
-    expect(spyPostMessage).toBeCalledWith({
-      msg: "nnsStopBalancesTimer",
-    });
+    postMessageMock.emit({
+      data: {
+        msg: "nnsSyncBalances",
+        data: {
+          balances: [
+            {
+              accountIdentifier: mockCkBTCMainAccount.identifier,
+              balance: 123456n,
+            },
+          ],
+        },
+      },
+    } as BalancesMessageEvent);
+
+    await waitFor(() => expect(spyReload).toHaveBeenCalledTimes(1));
   });
 });
